requirements.txt
asgiref==3.5.2
boto3==1.26.58
botocore==1.29.58
Django==4.0.6
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.5
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.14
whitenoise==6.3.0
